requirements.txt
opencv-python==4.5.5
numpy==1.23.5
matplotlib>=3.4.0
scikit-learn==1.1.1
scikit-image>=0.19.0
torch==2.0.1
torchvision==0.15.2
transformers>=4.20.0
tqdm>=4.62.0
pandas==1.5.3
Pillow>=9.0.0